Fix thread issuees

+++ ProxyManagement/ProxyThread.py
@@ -7,11 +7,12 @@ class ProxyThread(threading.Thread):
         self.stdout = subprocess.PIPE
         self.stdin = subprocess.PIPE
         self.proxy = proxy
+        self.p = None
         threading.Thread.__init__(self)
         self._stop = threading.Event()
     
     def run(self):
-        print('Proxy Name: {0}'.format(self.proxy))
+        print('Stop Proxy: {0}'.format(self.proxy))
         command = 'sudo openvpn --config /home/pi/Downloads/{0}.hideservers.net.ovpn'.format(self.proxy).split()
         self.p = subprocess.Popen(command,
                              shell=True,
@@ -19,4 +20,7 @@ class ProxyThread(threading.Thread):
                              stderr=subprocess.PIPE)
 	
     def stop(self):
-        self.p.kill()
+        import pdb; pdb.set_trace()
+        print('Start Proxy: {0}'.format(self.proxy))
+        if self.p is not None:
+            self.p.kill()
